Add new files about the data struct(DataStruct.h,DataStruct.cpp) and the storage of common params and configure(Tool.h,Tool.cpp)
The user can  interactively add points on the screen (click the left button to add point).The data struct class of point is called LECP_Point, which is defined in DataStruct.h file. All of the points added by user are stored in LECP_Doc class. The parameter points of LECP_Doc is represent the points user added. Every point added by user is unique,that is there is no same points in the points vector.